# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MobileMoney Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  userId: string;
  type: TransactionType;
  amount: number;
  currency: string;
  category: string;
  paymentMethod: string;
  /** @nullable */
  referenceNote?: string | null;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  currency: string;
  category: string;
  paymentMethod: string;
  /** @nullable */
  referenceNote?: string | null;
  date: string;
}

export type UpdateTransactionBodyType =
  (typeof UpdateTransactionBodyType)[keyof typeof UpdateTransactionBodyType];

export const UpdateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface UpdateTransactionBody {
  type?: UpdateTransactionBodyType;
  amount?: number;
  currency?: string;
  category?: string;
  paymentMethod?: string;
  /** @nullable */
  referenceNote?: string | null;
  date?: string;
}

export interface DashboardSummary {
  todayBalance: number;
  todayIncome: number;
  todayExpenses: number;
  weekIncome: number;
  weekExpenses: number;
  weekProfit: number;
  monthIncome: number;
  monthExpenses: number;
  monthProfit: number;
  totalTransactions: number;
  currency: string;
}

export interface DailyBreakdown {
  date: string;
  income: number;
  expenses: number;
  profit: number;
}

export interface CategoryBreakdown {
  category: string;
  type: string;
  total: number;
  count: number;
}

export interface PaymentMethodBreakdown {
  paymentMethod: string;
  total: number;
  count: number;
}

export interface InventoryItem {
  id: number;
  userId: string;
  name: string;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  notes?: string | null;
  updatedAt: string;
  createdAt: string;
}

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateInventoryItemBody {
  name?: string;
  quantity?: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ListTransactionsParams = {
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
  /**
   * @nullable
   */
  type?: string | null;
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  paymentMethod?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
};

export type GetSummaryByCategoryParams = {
  /**
   * week, month, or all
   * @nullable
   */
  period?: string | null;
};

export type GetSummaryByPaymentMethodParams = {
  /**
   * @nullable
   */
  period?: string | null;
};